refactor: put util function file back in general util
Relates #166
